feat: Update oh-my-posh and zsh configurations
This commit updates the oh-my-posh configuration to use the powerlevel
style and introduces a new environment variable, POSH_PATH_MAX_WIDTH,
which is set to the terminal width. This variable is used to limit the
maximum width of the path in the prompt. The zsh configuration is also
updated to define a new function, set_poshcontext, which sets the
POSH_PATH_MAX_WIDTH variable.

ref: https://github.com/JanDeDobbeleer/oh-my-posh/releases/tag/v21.12.0