NVIM Enhance Telescope file search with visual selection
This commit enhances the file search functionality in Neovim. It allows
the user to use the visually selected text as the default search text
when searching for files. This is achieved by modifying the keymap
settings in the 'neoscopes.lua' and 'telescope/init.lua' files. A new
utility function 'get_visual_selection_text' is also added in the
'utils/init.lua' file to get the visually selected text.