feat: Avoid using aliae because it increases the shell startup time
This commit updates the shell aliases and oh-my-posh setup in bash, zsh,
and PowerShell. It removes the dependency on aliae and directly sources
the aliases from dedicated files for each shell. It also ensures
oh-my-posh is installed and initialized correctly. The aliases and
functions related to git have been updated and some new ones have been
added.