feat(nvim): add nvim-rip-substitute plugin
This commit adds the nvim-rip-substitute plugin to the Neovim
configuration. This plugin provides a powerful substitute command
with a user-friendly interface. The keybindings for document and
workspace symbols in the telescope plugin have been updated to
avoid conflicts with the new plugin.

BREAKING CHANGE: The keybindings for document and workspace symbols
in the telescope plugin have been changed from <leader>fs and
<leader>fS to <leader>fSd and <leader>fSw respectively.