feat(lazy, statusline): Enable checker and add updates to statusline
In the `lazy` configuration, the checker has been enabled with the
options `notify` and `check_pinned` set to false. In the `statusline`
configuration, a new component has been added to display updates. The
update component is only displayed when there are updates available.
The color of the update component is set to `#ff9e64`.