feat(neovim): Enhance neoclip and whichkey plugins
This commit introduces new keybindings for the neoclip plugin, allowing
users to pull and push from the clipboard database. These keybindings
are intended to replace the continuous_sync option, which was disabled
due to performance issues. The keybindings are added under the 'Neoclip'
group in the whichkey plugin. This enhancement improves the usability of
the neoclip plugin.