feat(nvim-ufo): Enable auto foldcolumn and disable diagnostic
This commit enables the 'auto:4' setting for 'foldcolumn' in the
nvim-ufo plugin configuration. This will automatically adjust the
foldcolumn width up to a maximum of 4 characters.

Additionally, a diagnostic line has been disabled to prevent linting
warnings about the 'provider_selector' function. This function is
required for the plugin to function correctly.