feat(nvim): Add buffer-reopen plugin and update whichkey
This commit introduces the 'buffer-reopen.nvim' plugin to the Neovim
configuration. The plugin is set to trigger on a 'VeryLazy' event and
is bound to the '<C-S-t>' key combination in normal mode.

In addition, the 'whichkey.lua' file has been updated to reflect the
new plugin. The '<leader>b' key group, previously labeled 'Buffer',
has been renamed to 'Buffer Reopen' to better represent its new
functionality.