feat(nvim): Add vim-flog plugin and keybindings
This commit introduces the vim-flog plugin to the Neovim configuration.
It also adds keybindings for the plugin commands under the prefix
"<leader>gf". The plugin is set to lazy load to improve startup time.
The vim-fugitive plugin is added as a dependency as vim-flog relies on
it.

Additionally, the whichkey configuration is updated to include a new
group "Vim-Flog" under the prefix "<leader>gf".