refactor(komorebic.ahk): Refactor commands to use functions
The commands in the AutoHotkey script `komorebic.ahk` have been
refactored to use functions instead of direct `Run` calls. This makes
the code more readable and maintainable. The functions are defined in
`komorebic.lib.ahk` and imported into the main script. This change
affects a large number of lines but does not alter the functionality of
the script.