feat(tmux): Add tmux-floax plugin configuration
This commit introduces the tmux-floax plugin to the tmux configuration.
The plugin is bound to 'p' and 'C-P' for regular and menu usage
respectively. The floax window size is set to 80% of the parent window,
and the border and text colors are set to magenta and blue respectively.
The plugin is also configured to change the path when activated.