feat(terraform): add tfws-profile.sh for AWS profile export
Add `tfws-profile.sh` script to export AWS_PROFILE based on the current
Terraform workspace. The script maps workspace names to corresponding
AWS profiles and is intended to be used with `eval
"$(tfws-profile.sh)"`.

Update `.shared.sh` to provide a `tfws-profile` shell function and
`tfwsp` alias if the script is available in PATH.